chore: refactoring transformData
