+++ index.js
@@ -1,28 +1,7 @@
-
-import { validNrCpfCnpj, CSVRead, validatePrestationCalculation, brlFormatter, convertToDate } from "./utils.js"
-
+import { CSVRead } from "./utils.js"
+import { transformData } from "./transformData.js"
  
-const transformData = (results) => results.map(item => {
-    const { vlTotal, vlPresta, vlMora, nrCpfCnpj, qtPrestacoes, dtContrato, dtVctPre, ...rest } = item,
-        isValid = validatePrestationCalculation(vlTotal, qtPrestacoes, vlPresta);
-
-    return {
-        ...rest,
-        validCpfCnpj: {
-            ...validNrCpfCnpj(`${nrCpfCnpj}`)
-        },
-        vlTotal: brlFormatter.format(vlTotal),
-        vlPresta: brlFormatter.format(vlPresta),
-        vlMora: brlFormatter.format(vlMora),
-        vlPrestaInfo: isValid ? 'Cálculos de prestações consistentes.' : 'Cálculos de prestações inconsistentes.',
-        dtContrato: convertToDate(dtContrato),
-        dtVctPre: convertToDate(dtVctPre)
-    };
-})
-
-
-
-
+ 
 const run = async () => {
     try {
 
@@ -33,7 +12,7 @@ const run = async () => {
             (data) => results.push(data),
             () => {
 
-                const formatedResults = transformData(results);
+                const [formatedResults] = transformData(results);
 
                 console.log(formatedResults);
